refactor: move binomial sampling into its own module
@day01 and @YeungOnion both asked for this; the grouping is the one
@YeungOnion described.

`binomial.rs` was 1058 lines. It is now:

    binomial/mod.rs        556  the distribution: constructors, moments,
                                pmf/cdf, and their tests
    binomial/sampling.rs   824  BINV and BTPE, path selection, the algorithm
                                selection API, and their tests

Pure code movement -- no behaviour change. The 27 tests are the same 27, split
between the two modules; `--list` before and after gives an identical set of
names.

Two fixups carried along. `Binomial::sampler` and the `Distribution` impls
moved with the sampling code, since they are its entry points, and
`sample_unchecked` stays crate-internal as before. And the doc comment
describing the chi-square goodness-of-fit test had ended up above
`test_forced_algorithms_sample_correctly` when that test was inserted in the
previous commit; it is back on the test it describes.
